// libs/features/collection/billing/src/lib/shared/models/itemized-rejected-OH-wrapper.ts
import { RejectedOHDetails } from './rejected-OH-details';

/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class ItemizedRejectedOHWrapper {
  rejectedInjuryCount: number = undefined;
  amount: number = undefined;
  cases: RejectedOHDetails[] = [];
}