// libs/features/collection/billing/src/lib/shared/models/itemized-installment-wrapper.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class ItemizedInstallmentWrapper {
  currentInstallmentAmount: number = undefined;
  previouslyPaidAmount: number = undefined;
  remainingInstallmentAmount: number = undefined;
  totalInstallmentAmount: number = undefined;
  numberOfMonthsPaid: number = undefined;
  numberOfMonthsRemaining: number = undefined;
}